support config referHook
